backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pandas>=2.2.0
openai>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.10.0
langgraph>=1.0.0
